+++ packages/uniforms-bridge-json-schema/__tests__/JSONSchemaBridge.ts
@@ -459,7 +459,7 @@ describe('JSONSchemaBridge', () => {
 
     it('works with other errors', () => {
       expect(bridge.getErrorMessages('correct')).toEqual(['correct']);
-      expect(bridge.getErrorMessages(999999999)).toEqual([999999999]);
+      expect(bridge.getErrorMessages(999999999)).toEqual(['999999999']);
     });
 
     it('works with Error', () => {
@@ -492,6 +492,18 @@ describe('JSONSchemaBridge', () => {
           details: [{ instancePath: '.field', message: 'Ignore!' }],
         }),
       ).toEqual(['Ignore!']);
+    });
+
+    it('works with empty object', () => {
+      expect(bridge.getErrorMessages({})).toEqual([]);
+    });
+
+    it('works with empty messages', () => {
+      expect(
+        bridge.getErrorMessages({
+          details: [{ instancePath: '.field' }],
+        }),
+      ).toEqual(['']);
     });
   });
 
@@ -673,17 +685,7 @@ describe('JSONSchemaBridge', () => {
       });
     });
 
-    it('works with allowedValues from props', () => {
-      expect(
-        bridge.getProps('shippingAddress.type', { allowedValues: [1] }),
-      ).toEqual({
-        allowedValues: ['residential', 'business'],
-        label: 'Type',
-        required: true,
-      });
-    });
-
-    it('works with allowedValues from props', () => {
+    it('works with custom props', () => {
       expect(bridge.getProps('forcedRequired')).toEqual({
         label: 'Forced required',
         required: true,
@@ -702,67 +704,6 @@ describe('JSONSchemaBridge', () => {
       expect(bridge.getProps('withLabel')).toEqual({
         label: 'Example',
         required: false,
-      });
-    });
-
-    it('works with label (custom)', () => {
-      expect(bridge.getProps('dateOfBirth', { label: 'Death' })).toEqual({
-        label: 'Date of birth',
-        required: true,
-      });
-    });
-
-    it('works with label (true)', () => {
-      expect(bridge.getProps('dateOfBirth', { label: true })).toEqual({
-        label: 'Date of birth',
-        required: true,
-      });
-    });
-
-    it('works with property title as default label', () => {
-      expect(bridge.getProps('hasAJob', { label: true })).toEqual({
-        allowedValues: undefined,
-        label: 'Currently Employed',
-        options: undefined,
-        placeholder: undefined,
-        required: false,
-      });
-    });
-
-    it('works with placeholder (custom)', () => {
-      expect(bridge.getProps('email.work', { placeholder: 'Email' })).toEqual({
-        label: 'Work',
-        required: true,
-      });
-    });
-
-    it('works with placeholder (true)', () => {
-      expect(bridge.getProps('email.work', { placeholder: true })).toEqual({
-        label: 'Work',
-        required: true,
-      });
-    });
-
-    it('works with placeholder (falsy)', () => {
-      expect(bridge.getProps('email.work', { placeholder: null })).toEqual({
-        label: 'Work',
-        required: true,
-      });
-    });
-
-    it('works with placeholder (label falsy)', () => {
-      expect(
-        bridge.getProps('email.work', { label: null, placeholder: true }),
-      ).toEqual({
-        label: 'Work',
-        required: true,
-      });
-
-      expect(
-        bridge.getProps('email.work', { label: false, placeholder: true }),
-      ).toEqual({
-        label: 'Work',
-        required: true,
       });
     });
 
@@ -799,14 +740,6 @@ describe('JSONSchemaBridge', () => {
       expect(bridge.getProps('salary').allowedValues[1]).toBe('medium');
     });
 
-    it('works with options from props', () => {
-      const props = { options: { minimal: 4000, avarage: 8000 } };
-      expect(bridge.getProps('salary', props).transform('minimal')).toBe(4000);
-      expect(bridge.getProps('salary', props).transform('avarage')).toBe(8000);
-      expect(bridge.getProps('salary', props).allowedValues[0]).toBe('minimal');
-      expect(bridge.getProps('salary', props).allowedValues[1]).toBe('avarage');
-    });
-
     it('works with type', () => {
       expect(bridge.getProps('password')).toEqual({
         label: 'Password',
@@ -826,12 +759,10 @@ describe('JSONSchemaBridge', () => {
     });
 
     it('works with other props', () => {
-      expect(bridge.getProps('personalData.firstName', { x: 1, y: 1 })).toEqual(
-        {
-          label: 'First name',
-          required: false,
-        },
-      );
+      expect(bridge.getProps('personalData.firstName')).toEqual({
+        label: 'First name',
+        required: false,
+      });
     });
 
     it('works with allOf in items', () => {