+++ packages/uniforms/__suites__/render.tsx
@@ -1,16 +1,15 @@
 import { render as renderOnScreen, RenderResult } from '@testing-library/react';
-import React, { cloneElement, ReactElement } from 'react';
+import React, { ReactElement, cloneElement } from 'react';
 import SimpleSchema, { SimpleSchemaDefinition } from 'simpl-schema';
-import { BaseForm, context, Context, randomIds } from 'uniforms';
+import { BaseForm, Context, UnknownObject, context, randomIds } from 'uniforms';
 import { SimpleSchema2Bridge } from 'uniforms-bridge-simple-schema-2';
 
 const randomId = randomIds();
-
-export function render<P>(
+export function render<P, Model extends UnknownObject>(
   element: ReactElement<P>,
   schema?: SimpleSchemaDefinition,
-  contextValueExtension?: Partial<Context<unknown>>,
-  initialModel?: object,
+  contextValueExtension?: Partial<Context<Model>>,
+  model = {} as Model,
 ): RenderResult & { rerenderWithProps: (props: P) => void } {
   const renderResult = renderOnScreen(element, {
     wrapper({ children }) {
@@ -19,7 +18,7 @@ export function render<P>(
           changed: false,
           changedMap: {},
           error: null,
-          model: initialModel ?? {},
+          model,
           name: [],
           onChange() {},
           onSubmit() {},
@@ -37,7 +36,7 @@ export function render<P>(
             showInlineError: false,
             ...contextValueExtension?.state,
           },
-          formRef: {} as BaseForm<unknown>,
+          formRef: {} as BaseForm<UnknownObject>,
         };
         return (
           <context.Provider value={contextValue}>{children}</context.Provider>